chore: update webpack config
