+++ extensions/github1s/webpack.config.js
