+++ package.json
@@ -11,11 +11,14 @@
     "@typescript-eslint/eslint-plugin": "^5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
     "chokidar": "^3.5.3",
+    "clean-css": "^5.3.2",
     "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.7.3",
     "eslint": "^8.25.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-jsdoc": "^39.3.14",
     "eslint-plugin-prettier": "^4.2.1",
+    "file-loader": "^6.2.0",
     "fs-extra": "^10.1.0",
     "generate-file-webpack-plugin": "^1.0.1",
     "html-webpack-plugin": "^5.5.0",
@@ -26,8 +29,11 @@
     "prettier": "^2.7.1",
     "serve-handler": "^6.1.3",
     "start-server-and-test": "^1.14.0",
+    "style-loader": "^3.3.1",
+    "ts-loader": "^9.4.2",
     "typescript": "^4.7.3",
-    "webpack": "^5.73.0",
+    "uglify-js": "^3.17.4",
+    "webpack": "^5.75.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.9.2",
     "yalc": "^1.0.0-pre.53"
@@ -35,20 +41,17 @@
   "scripts": {
     "preinstall": "./scripts/preinstall.sh",
     "postinstall": "husky install && ./scripts/postinstall.sh",
-    "clean-build": "./scripts/clean-build.sh",
-    "clean": "./scripts/clean.sh",
+    "dev": "yarn watch",
     "build": "./scripts/build.sh",
+    "clean": "./scripts/clean.sh",
     "yalc": "yalc publish vscode-web && yalc add -D @github1s/vscode-web",
     "build:vscode": "cd vscode-web && yarn build",
-    "vercel-build": "USE_HASH_HTML=true yarn build && yarn hash && rm -rf public && mv dist public",
-    "hash": "./scripts/hash.sh",
-    "package": "yarn webpack",
     "watch": "rm -rf dist && run-p watch:*",
-    "watch-with-vscode": "DEV_VSCODE=true run-p watch watch-vscode",
     "watch-vscode": "cd vscode-web && yarn watch",
-    "watch:dev-server": "node server.js",
-    "webpack": "webpack --config webpack.config.js",
+    "watch-with-vscode": "DEV_VSCODE=true run-p watch watch-vscode",
+    "watch:dev-server": "webpack serve --mode=development",
     "watch:github1s-extension": "cd extensions/github1s && yarn dev",
+    "webpack": "webpack --config webpack.config.js",
     "format": "prettier --write .",
     "eslint": "eslint --ext .ts --ext .js .",
     "test:ci": "start-test watch:dev-server 5000 test",
