--- resources/index-dev-vscode.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="utf-8" />
-		<!-- Disable pinch zooming -->
-		<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no" />
-		<!-- Workbench Icon/Manifest/CSS -->
-		<link rel="icon" media="(prefers-color-scheme:light)" href="/favicon-light.svg" type="image/svg+xml" />
-		<link rel="icon" media="(prefers-color-scheme:dark)" href="/favicon-dark.svg" type="image/svg+xml" />
-		<link rel="manifest" href="/manifest.json" />
-		<title>GitHub1s</title>
-		<script><%= titleScript %></script>
-	</head>
-
-	<body aria-label="">
-		<noscript title="No JavaScript Support">
-			<h1>You need to enable JavaScript to run this app.</h1>
-		</noscript>
-	</body>
-
-	<script src="/static/vscode/vs/loader.js"></script>
-	<script src="/static/vscode/vs/webPackagePaths.js"></script>
-	<script src="/static/config/extensions.js"></script>
-	<script src="/static/config/initialize.js"></script>
-	<script>
-		require(['vs/code/browser/workbench/workbench'], function () {});
-	</script>
-	<script src="https://unpkg.com/favicon-switcher@1.2.2/dist/index.js" crossorigin="anonymous" type="application/javascript" async="async"></script>
-</html>
