--- resources/index-hash.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="utf-8" />
-		<!-- Disable pinch zooming -->
-		<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no" />
-		<!-- Workbench Icon/Manifest/CSS -->
-		<link rel="icon" media="(prefers-color-scheme:light)" href="/favicon-light.svg" type="image/svg+xml">
-		<link rel="icon" media="(prefers-color-scheme:dark)" href="/favicon-dark.svg" type="image/svg+xml">
-		<link rel="preload" href="/static/{STATIC_HASH_CODE}/vscode/vs/workbench/workbench.web.main.css" as="style">
-		<link rel="manifest" href="/manifest.json" />
-		<title>GitHub1s</title>
-		<style>#load-spinner{position:absolute;top:0;left:0;right:0;bottom:0;display:flex;flex-direction:column;align-items:center;justify-content:center;background-color:#1e1e1e}.lds-grid{display:inline-block;position:relative;width:80px;height:80px}.lds-grid div{position:absolute;width:16px;height:16px;border-radius:50%;background:#fff;animation:lds-grid 1.2s linear infinite}.lds-grid div:nth-child(1){top:8px;left:8px;animation-delay:0s}.lds-grid div:nth-child(2){top:8px;left:32px;animation-delay:-.4s}.lds-grid div:nth-child(3){top:8px;left:56px;animation-delay:-.8s}.lds-grid div:nth-child(4){top:32px;left:8px;animation-delay:-.4s}.lds-grid div:nth-child(5){top:32px;left:32px;animation-delay:-.8s}.lds-grid div:nth-child(6){top:32px;left:56px;animation-delay:-1.2s}.lds-grid div:nth-child(7){top:56px;left:8px;animation-delay:-.8s}.lds-grid div:nth-child(8){top:56px;left:32px;animation-delay:-1.2s}.lds-grid div:nth-child(9){top:56px;left:56px;animation-delay:-1.6s}@keyframes lds-grid{0%,100%{opacity:1}50%{opacity:.5}}</style>
-		<script><%= titleScript %></script>
-		<script>(function(){var mid='G-D3LJNF4DN0';if(window.location.hostname.match(/\.?gitlab1s\.com$/i)){mid='G-1F70ST6944';}else if(window.location.hostname.match(/\.?bitbucket1s\.org$/i)){mid='G-SPWSR3V6YC';}else if(window.location.hostname.match(/\.?npmjs1s\.com$/i)){mid='G-VF4VCXYFKV';};window.dataLayer=window.dataLayer||[];function gtag(){dataLayer.push(arguments);};gtag('js',new Date());gtag('config',mid);window.addEventListener('load',function(){var element=window.document.createElement('script');element.src='https://www.googletagmanager.com/gtag/js?id='+mid;window.document.body.append(element);});})();</script>
-	</head>
-
-	<body aria-label="">
-		<noscript title="No JavaScript Support">
-			<h1>You need to enable JavaScript to run this app.</h1>
-		</noscript>
-		<div id="load-spinner" aria-label="loading">
-			<div class="lds-grid">
-				<div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div>
-			</div>
-		</div>
-	</body>
-
-	<script>window.staticHashCode = '{STATIC_HASH_CODE}';</script>
-	<script src="/static/{STATIC_HASH_CODE}/vscode/vs/loader.js"></script>
-	<script src="/static/{STATIC_HASH_CODE}/vscode/vs/webPackagePaths.js"></script>
-	<script src="/static/{STATIC_HASH_CODE}/config/extensions.js"></script>
-	<script src="/static/{STATIC_HASH_CODE}/config/initialize.js"></script>
-	<script src="/static/{STATIC_HASH_CODE}/vscode/vs/workbench/workbench.web.main.nls.js"></script>
-	<script src="/static/{STATIC_HASH_CODE}/vscode/vs/workbench/workbench.web.main.js"></script>
-	<script src="/static/{STATIC_HASH_CODE}/vscode/vs/code/browser/workbench/workbench.js"></script>
-	<script src="https://unpkg.com/favicon-switcher@1.2.2/dist/index.js" crossorigin="anonymous" type="application/javascript" async="async"></script>
-</html>
