+++ scripts/clean.sh
@@ -6,9 +6,10 @@ APP_ROOT=$(pwd)
 
 # execute all necessary tasks
 function main() {
-	./scripts/clean-build.sh
 	cd ${APP_ROOT}
-	rm -rf node_modules lib extensions/github1s/node_modules
+	rm -rf dist extensions/**/dist && echo "clean dist"
+	rm -rf vscode-web/lib vscode-web/node_modules && echo "clean vscode-web"
+	rm -rf node_modules extensions/**/node_modules && echo "clean node_modules"
 	echo "clean all"
 }
