+++ src/global.d.ts
@@ -1,0 +1,6 @@
+declare module '*.svg' {
+	const content: string;
+	export default content;
+}
+
+declare const STATIC_HASH: string;
