+++ src/notification.css
@@ -1,0 +1,85 @@
+.github1s-notification {
+  display: block;
+  position: fixed;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  height: 60px;
+  z-index: 999;
+  display: flex;
+  box-shadow: 1px 1px 5px 3px #1e1e1e;
+  background: rgba(0, 0, 0, .8);
+  font-size: 14px
+}
+
+.github1s-notification .notification-main {
+  flex: 1;
+  padding: 0 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  overflow: hidden
+}
+
+.github1s-notification .notification-main .notification-title {
+  color: #ffe58f;
+  font-size: 16px;
+  margin-bottom: 2px;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis
+}
+
+.github1s-notification .notification-main .notification-content {
+  color: #ccc;
+  font-size: 13px;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis
+}
+
+.github1s-notification .notification-main .notification-link {
+  color: #3794ff;
+  text-decoration: none
+}
+
+.github1s-notification .notification-main .notification-link:focus {
+  outline: 1px solid #007fd4;
+  outline-offset: -1px
+}
+
+.github1s-notification .notification-footer {
+  padding-right: 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  min-width: 140px
+}
+
+.github1s-notification .notification-footer .notification-confirm-button {
+  border: none;
+  width: 100%;
+  height: 26px;
+  margin-bottom: 2px;
+  text-align: center;
+  outline: 1px solid transparent;
+  outline-offset: 2px !important;
+  color: #fff;
+  background-color: #0e639c;
+  cursor: pointer
+}
+
+.github1s-notification .notification-footer .notification-confirm-button:hover {
+  background-color: #17b
+}
+
+.github1s-notification .notification-footer .notification-confirm-button:focus {
+  outline-color: #007fd4
+}
+
+.github1s-notification .notification-footer .notification-show-me-again {
+  color: #ccc;
+  font-size: 12px;
+  display: flex;
+  align-items: center
+}
