+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
+	"include": ["./src/"],
 	"compilerOptions": {
 		"module": "amd",
 		"moduleResolution": "node",
