+++ webpack.config.js
@@ -1,80 +1,18 @@
 const path = require('path');
+const fs = require('fs-extra');
+const cp = require('child_process');
+const webpack = require('webpack');
+const CleanCSS = require('clean-css');
+const UglifyJS = require('uglify-js');
 const CopyPlugin = require('copy-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const generate = require('generate-file-webpack-plugin');
-const fs = require('fs-extra');
+const packUtils = require('./scripts/webpack.js');
 
-const APP_ROOT = path.join(__dirname, '.');
-
+const GIT_COMMIT_ID = cp.execSync('git rev-parse HEAD').toString().trim();
+const STATIC_HASH = GIT_COMMIT_ID.padStart(7, '0').slice(0, 7);
 const devVscode = !!process.env.DEV_VSCODE;
-const useHashHtml = !!process.env.USE_HASH_HTML;
-
-const isWebExtension = (manifest) => {
-	if (Boolean(manifest.browser)) {
-		return true;
-	}
-	if (Boolean(manifest.main)) {
-		return false;
-	}
-	// neither browser nor main
-	if (typeof manifest.extensionKind !== 'undefined') {
-		const extensionKind = Array.isArray(manifest.extensionKind) ? manifest.extensionKind : [manifest.extensionKind];
-		if (extensionKind.indexOf('web') >= 0) {
-			return true;
-		}
-	}
-	if (typeof manifest.contributes !== 'undefined') {
-		for (const id of ['debuggers', 'terminal', 'typescriptServerPlugins']) {
-			if (manifest.contributes.hasOwnProperty(id)) {
-				return false;
-			}
-		}
-	}
-	return true;
-};
-
-const getExtensionData = (absolutePath) => {
-	try {
-		const packageJSONPath = path.join(absolutePath, 'package.json');
-		if (!fs.existsSync(packageJSONPath)) {
-			return null;
-		}
-		const packageJSON = JSON.parse(fs.readFileSync(packageJSONPath).toString('utf8'));
-		if (!isWebExtension(packageJSON)) {
-			return null;
-		}
-		const children = fs.readdirSync(absolutePath);
-		const packageNLSPath = children.filter((child) => child === 'package.nls.json')[0];
-		const packageNLS = packageNLSPath
-			? JSON.parse(fs.readFileSync(path.join(absolutePath, packageNLSPath)).toString())
-			: undefined;
-		const readme = children.filter((child) => /^readme(\.txt|\.md|)$/i.test(child))[0];
-		const changelog = children.filter((child) => /^changelog(\.txt|\.md|)$/i.test(child))[0];
-		const extensionFolder = path.basename(absolutePath);
-
-		return {
-			extensionPath: extensionFolder,
-			packageJSON,
-			packageNLS,
-			readmePath: readme ? path.join(extensionFolder, readme) : undefined,
-			changelogPath: changelog ? path.join(extensionFolder, changelog) : undefined,
-		};
-	} catch {
-		return null;
-	}
-};
-
-const scanVSCodeExtensions = () => {
-	const extensionsPath = path.join(APP_ROOT, 'node_modules/@github1s/vscode-web/dist/extensions');
-	const extensionFolders = fs.existsSync(extensionsPath) ? fs.readdirSync(extensionsPath) : [];
-
-	return extensionFolders.map((item) => getExtensionData(path.join(extensionsPath, item))).filter(Boolean);
-};
-
-const scanGitHub1sExtensions = () => {
-	const extensions = fs.readdirSync(path.join(APP_ROOT, 'extensions'));
-	return extensions.map((item) => getExtensionData(path.join(APP_ROOT, 'extensions', item))).filter(Boolean);
-};
+const skipMinified = { info: { minimized: true } };
 
 const VSCODE_NODE_MODULES = [
 	'@vscode/iconv-lite-umd',
@@ -87,104 +25,98 @@ const VSCODE_NODE_MODULES = [
 	'xterm-addon-search',
 	'xterm-addon-unicode11',
 	'xterm-addon-webgl',
-]
-	.map((x) => `vscode-web/node_modules/${x}/**`)
-	.map((from) => ({
-		from,
-		globOptions: {
-			dot: true,
-			// ignore: ["**/node_modules/**/node_modules/**"]
-		},
-		to({ context, absoluteFilename }) {
-			const relativePath = path.relative(context, absoluteFilename);
-			const relativeDir = path.dirname(relativePath.replace('vscode-web/node_modules/', ''));
-			return `static/node_modules/${relativeDir}/[name][ext]`;
-		},
-	}));
-
-module.exports = {
-	mode: 'development',
-	entry: './resources/manifest.json',
-	plugins: [
-		new CopyPlugin({
-			patterns: [
-				{ from: 'resources/favicon*', to: '[name][ext]' },
-				{ from: 'resources/manifest.json', to: '[name][ext]' },
-				{ from: 'resources/robots.txt', to: '[name][ext]' },
-				{
-					from: 'node_modules/@github1s/vscode-web/dist/extensions/**',
-					to({ context, absoluteFilename }) {
-						const relativePath = path.relative(context, absoluteFilename);
-						const relativeDir = path.dirname(
-							relativePath.replace('node_modules/@github1s/vscode-web/dist/extensions/', '')
-						);
-						return `static/extensions/${relativeDir}/[name][ext]`;
-					},
-				},
-				!devVscode && {
-					from: 'node_modules/@github1s/vscode-web/dist/vscode/',
-					to({ context, absoluteFilename }) {
-						const relativePath = path.relative(context, absoluteFilename);
-						const relativeDir = path.dirname(
-							relativePath.replace('node_modules/@github1s/vscode-web/dist/vscode/', '')
-						);
-						return `static/vscode/${relativeDir}/[name][ext]`;
-					},
-				},
-				{
-					from: 'extensions/**/*',
-					to: 'static',
-					globOptions: {
-						dot: true,
-						ignore: ['**/node_modules/**'],
-					},
-				},
-				{
-					from: 'resources/{initialize.js,github.svg,gitlab.svg,bitbucket.svg,npm.svg}',
-					to: 'static/config/[name][ext]',
-				},
-				...VSCODE_NODE_MODULES,
-			].filter(Boolean),
-		}),
-		new HtmlWebpackPlugin({
-			templateParameters: {
-				titleScript: fs.readFileSync('resources/titleScript.js'),
-			},
-			template: useHashHtml
-				? 'resources/index-hash.html'
-				: devVscode
-				? 'resources/index-dev-vscode.html'
-				: 'resources/index.html',
-			inject: false,
-		}),
-		generate({
-			file: 'static/config/extensions.js',
-			content: () => {
-				const vscodeExtensions = devVscode ? scanVSCodeExtensions() : [];
-				const extensions = [...vscodeExtensions, ...scanGitHub1sExtensions()];
-				return `window.github1sExtensions = ${JSON.stringify(extensions)};`;
-			},
-		}),
-	],
-	devServer: {
-		static: {
-			directory: path.join(__dirname, 'dist'),
-		},
-		allowedHosts: 'all',
-		client: {
-			progress: true,
-		},
-		historyApiFallback: {
-			rewrites: [{ from: /./, to: '/index.html' }],
-		},
-		devMiddleware: {
-			writeToDisk: true,
-		},
-		proxy: {
-			'/api/vscode-unpkg': {
-				target: 'http://localhost:5001',
-			},
-		},
-		port: 5000,
+].map((pkg) => ({
+	from: `vscode-web/node_modules/${pkg}/**`,
+	globOptions: { dot: true },
+	to({ context, absoluteFilename }) {
+		const relativePath = path.relative(context, absoluteFilename);
+		const relativeDir = path.dirname(relativePath.replace('vscode-web/node_modules/', ''));
+		return `static-${STATIC_HASH}/node_modules/${relativeDir}/[name][ext]`;
 	},
+	...skipMinified,
+}));
+
+module.exports = (env, argv) => {
+	const devMode = argv.mode === 'development';
+	const minifyCSS = (code) => (devMode ? code : new CleanCSS().minify(code).styles);
+	const minifyJS = (code) => (devMode ? code : UglifyJS.minify(code).code);
+
+	return {
+		mode: env.mode || 'production',
+		entry: path.resolve(__dirname, 'src/index.ts'),
+		output: { filename: `static-${STATIC_HASH}/config/bootstrap.js` },
+		resolve: { extensions: ['.js', '.ts'] },
+		module: {
+			rules: [
+				{ test: /\.tsx?$/, use: 'ts-loader' },
+				{ test: /\.css?$/, use: ['style-loader', 'css-loader'] },
+				{ test: /\.svg$/, use: 'file-loader' },
+			],
+		},
+		plugins: [
+			new CopyPlugin({
+				patterns: [
+					{ from: 'public/favicon*', to: '[name][ext]' },
+					{ from: 'public/manifest.json', to: '[name][ext]' },
+					{ from: 'public/robots.txt', to: '[name][ext]' },
+					{
+						from: 'extensions',
+						to: `static-${STATIC_HASH}/extensions`,
+						globOptions: { dot: true, ignore: ['**/node_modules/**'] },
+						...skipMinified,
+					},
+					!devVscode && {
+						from: 'node_modules/@github1s/vscode-web/dist/vscode',
+						to: `static-${STATIC_HASH}/vscode`,
+						...skipMinified,
+					},
+					!devVscode && {
+						from: 'node_modules/@github1s/vscode-web/dist/extensions',
+						to: `static-${STATIC_HASH}/extensions`,
+						...skipMinified,
+					},
+					...VSCODE_NODE_MODULES,
+				].filter(Boolean),
+			}),
+			new HtmlWebpackPlugin({
+				inject: false,
+				template: 'public/index.html',
+				templateParameters: {
+					devVscode: devVscode,
+					staticHash: STATIC_HASH,
+					spinnerStyle: minifyCSS(fs.readFileSync('./public/spinner.css').toString()),
+					pageTitleScript: minifyJS(fs.readFileSync('./public/page-title.js').toString()),
+					analyticsScript: devMode ? '' : minifyJS(fs.readFileSync('./public/analytics.js').toString()),
+				},
+			}),
+			new webpack.DefinePlugin({
+				STATIC_HASH: JSON.stringify(STATIC_HASH),
+			}),
+			generate({
+				file: `static-${STATIC_HASH}/config/extensions.js`,
+				content: packUtils.createExtensionsContent(devVscode),
+			}),
+		],
+		performance: false,
+		devServer: {
+			port: 5000,
+			liveReload: false,
+			allowedHosts: 'all',
+			static: {
+				directory: path.join(__dirname, 'dist'),
+			},
+			client: {
+				progress: true,
+			},
+			historyApiFallback: {
+				rewrites: [{ from: /./, to: '/index.html' }],
+			},
+			devMiddleware: {
+				writeToDisk: true,
+			},
+			proxy: {
+				'/api/vscode-unpkg': packUtils.createVSCodeUnpkgProxy(),
+			},
+		},
+	};
 };
